+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Remove usages of the `Quiet` switch from usages of `Set-CRegistryKeyValue`. That
 
 * Replaced verbose-level messages with information-level messages in `Install-CRegistryKey`, `Remove-CRegistryKeyValue`,
 and `Set-CRegistryKeyValue` when saving changes.
+* `Set-CRegistryKeyValue` accepts `$null` as the value of a multi-line string, which sets the value to an empty list.
 
 ### Fixed